Fixing tests and adjust for server sent events
