+++ package.json
@@ -33,7 +33,9 @@
     "@types/react": "^16.9.17",
     "@types/react-dom": "^16.9.4",
     "@types/socket.io-client": "^1.4.32",
+    "cors": "^2.8.5",
     "cypress": "^3.8.1",
+    "express": "^4.17.1",
     "parcel-bundler": "^1.12.4",
     "prettier": "^1.19.1",
     "react": "^16.12.0",
